debug

+++ src/inference/inference.cuh
@@ -24,7 +24,7 @@ void copy_fp16_tensor(Tensor *Y, Tensor *X);
 
 void compute_layer_norm(Tensor *RMSNorm, Tensor *X, __half *d_gcache);
 
-__global__ void kernel_compute_rms_norm(__half *X_tensor, __half *RMSNorm_tensor, __half *d_gcache);
+__global__ void kernel_compute_rms_norm(__half *X_tensor, __half *RMSNorm_tensor, float *d_gcache);
 
 /* ******************************* Attention Computation ******************************* */
 void _create_intermediary_attention_tensor(Tensor *Attention_Tensor, Tensor *Linear);
